# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for املاک AI - Tehran real estate platform
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type EstimatePriceBodyPropertyType =
  (typeof EstimatePriceBodyPropertyType)[keyof typeof EstimatePriceBodyPropertyType];

export const EstimatePriceBodyPropertyType = {
  apartment: "apartment",
  villa: "villa",
  office: "office",
  commercial: "commercial",
} as const;

export interface EstimatePriceBody {
  /** Property area in square meters */
  area: number;
  /** Tehran neighborhood (e.g. پونک, الهیه, شریعتی) */
  neighborhood: string;
  /** Year of construction (Shamsi calendar) */
  yearBuilt?: number;
  /** Number of bedrooms */
  rooms?: number;
  /** Floor number */
  floor?: number;
  /** Total floors in building */
  totalFloors?: number;
  hasParking?: boolean;
  hasElevator?: boolean;
  hasStorage?: boolean;
  propertyType: EstimatePriceBodyPropertyType;
}

export type EstimatePriceResponsePriceRange = {
  min: number;
  max: number;
};

export type EstimatePriceResponseConfidence =
  (typeof EstimatePriceResponseConfidence)[keyof typeof EstimatePriceResponseConfidence];

export const EstimatePriceResponseConfidence = {
  high: "high",
  medium: "medium",
  low: "low",
} as const;

export interface EstimatePriceResponse {
  /** Estimated total price in Tomans */
  estimatedPrice: number;
  /** Price per square meter in Tomans */
  pricePerMeter: number;
  priceRange: EstimatePriceResponsePriceRange;
  confidence: EstimatePriceResponseConfidence;
  /** AI analysis text in Persian */
  analysis: string;
  /** Market trend description in Persian */
  marketTrend: string;
}

export type GenerateAdBodyPropertyType =
  (typeof GenerateAdBodyPropertyType)[keyof typeof GenerateAdBodyPropertyType];

export const GenerateAdBodyPropertyType = {
  apartment: "apartment",
  villa: "villa",
  office: "office",
  commercial: "commercial",
} as const;

export type GenerateAdBodyTransactionType =
  (typeof GenerateAdBodyTransactionType)[keyof typeof GenerateAdBodyTransactionType];

export const GenerateAdBodyTransactionType = {
  sale: "sale",
  rent: "rent",
  mortgage: "mortgage",
} as const;

/**
 * Target platform for the ad
 */
export type GenerateAdBodyPlatform =
  (typeof GenerateAdBodyPlatform)[keyof typeof GenerateAdBodyPlatform];

export const GenerateAdBodyPlatform = {
  divar: "divar",
  sheypoor: "sheypoor",
  general: "general",
} as const;

export interface GenerateAdBody {
  propertyType: GenerateAdBodyPropertyType;
  transactionType: GenerateAdBodyTransactionType;
  area: number;
  rooms?: number;
  neighborhood: string;
  floor?: number;
  /** List of features (e.g. نورگیر, آفتابگیر, نوساز) */
  features?: string[];
  price?: number;
  contactInfo?: string;
  /** Target platform for the ad */
  platform: GenerateAdBodyPlatform;
}

export interface GenerateAdResponse {
  /** Ad title in Persian */
  title: string;
  /** Full ad description in Persian */
  description: string;
  /** Suggested tags for the listing */
  tags: string[];
}

export interface SearchPropertyBody {
  /** Natural language search query in Persian */
  query: string;
}

export type SearchPropertyResponseCriteria = {
  neighborhood?: string;
  minArea?: number;
  maxArea?: number;
  rooms?: number;
  propertyType?: string;
  maxPrice?: number;
  features?: string[];
};

export interface SearchPropertyResponse {
  /** AI interpretation of the search query in Persian */
  interpretation: string;
  criteria: SearchPropertyResponseCriteria;
  /** Search suggestions and alternatives in Persian */
  suggestions: string[];
}

export interface OpenaiConversation {
  id: number;
  title: string;
  createdAt: string;
}

export interface OpenaiMessage {
  id: number;
  conversationId: number;
  role: string;
  content: string;
  createdAt: string;
}

export interface CreateOpenaiConversationBody {
  title: string;
}

export interface SendOpenaiMessageBody {
  content: string;
}

export interface OpenaiConversationWithMessages {
  id: number;
  title: string;
  createdAt: string;
  messages: OpenaiMessage[];
}

export interface OpenaiError {
  error: string;
}