# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification for املاک AI - Tehran real estate platform
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: openai
    description: OpenAI AI chat, image, and audio operations
  - name: property
    description: Property AI tools - price estimation, ad generation, search
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /property/estimate-price:
    post:
      operationId: estimatePrice
      tags: [property]
      summary: Estimate property price using AI
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EstimatePriceBody"
      responses:
        "200":
          description: Price estimation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EstimatePriceResponse"

  /property/generate-ad:
    post:
      operationId: generateAd
      tags: [property]
      summary: Generate property ad text using AI
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateAdBody"
      responses:
        "200":
          description: Generated ad text
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenerateAdResponse"

  /property/search:
    post:
      operationId: searchProperty
      tags: [property]
      summary: AI-powered natural language property search
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SearchPropertyBody"
      responses:
        "200":
          description: Search results with AI analysis
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SearchPropertyResponse"

  /openai/conversations:
    get:
      operationId: listOpenaiConversations
      tags: [openai]
      summary: List all conversations
      responses:
        "200":
          description: List of conversations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OpenaiConversation"
    post:
      operationId: createOpenaiConversation
      tags: [openai]
      summary: Create a new conversation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOpenaiConversationBody"
      responses:
        "201":
          description: Created conversation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiConversation"

  /openai/conversations/{id}:
    get:
      operationId: getOpenaiConversation
      tags: [openai]
      summary: Get conversation with messages
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Conversation with messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiConversationWithMessages"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiError"
    delete:
      operationId: deleteOpenaiConversation
      tags: [openai]
      summary: Delete a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiError"

  /openai/conversations/{id}/messages:
    get:
      operationId: listOpenaiMessages
      tags: [openai]
      summary: List messages in a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OpenaiMessage"
    post:
      operationId: sendOpenaiMessage
      tags: [openai]
      summary: Send a text message and receive a streaming text response
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendOpenaiMessageBody"
      responses:
        "200":
          description: SSE stream of assistant text chunks
          content:
            text/event-stream: {}

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    EstimatePriceBody:
      type: object
      properties:
        area:
          type: number
          description: Property area in square meters
        neighborhood:
          type: string
          description: Tehran neighborhood (e.g. پونک, الهیه, شریعتی)
        yearBuilt:
          type: integer
          description: Year of construction (Shamsi calendar)
        rooms:
          type: integer
          description: Number of bedrooms
        floor:
          type: integer
          description: Floor number
        totalFloors:
          type: integer
          description: Total floors in building
        hasParking:
          type: boolean
        hasElevator:
          type: boolean
        hasStorage:
          type: boolean
        propertyType:
          type: string
          enum: [apartment, villa, office, commercial]
      required:
        - area
        - neighborhood
        - propertyType

    EstimatePriceResponse:
      type: object
      properties:
        estimatedPrice:
          type: number
          description: Estimated total price in Tomans
        pricePerMeter:
          type: number
          description: Price per square meter in Tomans
        priceRange:
          type: object
          properties:
            min:
              type: number
            max:
              type: number
          required: [min, max]
        confidence:
          type: string
          enum: [high, medium, low]
        analysis:
          type: string
          description: AI analysis text in Persian
        marketTrend:
          type: string
          description: Market trend description in Persian
      required:
        - estimatedPrice
        - pricePerMeter
        - priceRange
        - confidence
        - analysis
        - marketTrend

    GenerateAdBody:
      type: object
      properties:
        propertyType:
          type: string
          enum: [apartment, villa, office, commercial]
        transactionType:
          type: string
          enum: [sale, rent, mortgage]
        area:
          type: number
        rooms:
          type: integer
        neighborhood:
          type: string
        floor:
          type: integer
        features:
          type: array
          items:
            type: string
          description: List of features (e.g. نورگیر, آفتابگیر, نوساز)
        price:
          type: number
        contactInfo:
          type: string
        platform:
          type: string
          enum: [divar, sheypoor, general]
          description: Target platform for the ad
      required:
        - propertyType
        - transactionType
        - area
        - neighborhood
        - platform

    GenerateAdResponse:
      type: object
      properties:
        title:
          type: string
          description: Ad title in Persian
        description:
          type: string
          description: Full ad description in Persian
        tags:
          type: array
          items:
            type: string
          description: Suggested tags for the listing
      required:
        - title
        - description
        - tags

    SearchPropertyBody:
      type: object
      properties:
        query:
          type: string
          description: Natural language search query in Persian
      required:
        - query

    SearchPropertyResponse:
      type: object
      properties:
        interpretation:
          type: string
          description: AI interpretation of the search query in Persian
        criteria:
          type: object
          properties:
            neighborhood:
              type: string
            minArea:
              type: number
            maxArea:
              type: number
            rooms:
              type: integer
            propertyType:
              type: string
            maxPrice:
              type: number
            features:
              type: array
              items:
                type: string
        suggestions:
          type: array
          items:
            type: string
          description: Search suggestions and alternatives in Persian
      required:
        - interpretation
        - criteria
        - suggestions

    OpenaiConversation:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - createdAt

    OpenaiMessage:
      type: object
      properties:
        id:
          type: integer
        conversationId:
          type: integer
        role:
          type: string
        content:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - conversationId
        - role
        - content
        - createdAt

    CreateOpenaiConversationBody:
      type: object
      properties:
        title:
          type: string
      required:
        - title

    SendOpenaiMessageBody:
      type: object
      properties:
        content:
          type: string
      required:
        - content

    OpenaiConversationWithMessages:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        createdAt:
          type: string
          format: date-time
        messages:
          type: array
          items:
            $ref: "#/components/schemas/OpenaiMessage"
      required:
        - id
        - title
        - createdAt
        - messages

    OpenaiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error