# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for املاک AI - Tehran real estate platform
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Estimate property price using AI
 */
export const EstimatePriceBody = zod.object({
  area: zod.number().describe("Property area in square meters"),
  neighborhood: zod
    .string()
    .describe("Tehran neighborhood (e.g. پونک, الهیه, شریعتی)"),
  yearBuilt: zod
    .number()
    .optional()
    .describe("Year of construction (Shamsi calendar)"),
  rooms: zod.number().optional().describe("Number of bedrooms"),
  floor: zod.number().optional().describe("Floor number"),
  totalFloors: zod.number().optional().describe("Total floors in building"),
  hasParking: zod.boolean().optional(),
  hasElevator: zod.boolean().optional(),
  hasStorage: zod.boolean().optional(),
  propertyType: zod.enum(["apartment", "villa", "office", "commercial"]),
});

export const EstimatePriceResponse = zod.object({
  estimatedPrice: zod.number().describe("Estimated total price in Tomans"),
  pricePerMeter: zod.number().describe("Price per square meter in Tomans"),
  priceRange: zod.object({
    min: zod.number(),
    max: zod.number(),
  }),
  confidence: zod.enum(["high", "medium", "low"]),
  analysis: zod.string().describe("AI analysis text in Persian"),
  marketTrend: zod.string().describe("Market trend description in Persian"),
});

/**
 * @summary Generate property ad text using AI
 */
export const GenerateAdBody = zod.object({
  propertyType: zod.enum(["apartment", "villa", "office", "commercial"]),
  transactionType: zod.enum(["sale", "rent", "mortgage"]),
  area: zod.number(),
  rooms: zod.number().optional(),
  neighborhood: zod.string(),
  floor: zod.number().optional(),
  features: zod
    .array(zod.string())
    .optional()
    .describe("List of features (e.g. نورگیر, آفتابگیر, نوساز)"),
  price: zod.number().optional(),
  contactInfo: zod.string().optional(),
  platform: zod
    .enum(["divar", "sheypoor", "general"])
    .describe("Target platform for the ad"),
});

export const GenerateAdResponse = zod.object({
  title: zod.string().describe("Ad title in Persian"),
  description: zod.string().describe("Full ad description in Persian"),
  tags: zod.array(zod.string()).describe("Suggested tags for the listing"),
});

/**
 * @summary AI-powered natural language property search
 */
export const SearchPropertyBody = zod.object({
  query: zod.string().describe("Natural language search query in Persian"),
});

export const SearchPropertyResponse = zod.object({
  interpretation: zod
    .string()
    .describe("AI interpretation of the search query in Persian"),
  criteria: zod.object({
    neighborhood: zod.string().optional(),
    minArea: zod.number().optional(),
    maxArea: zod.number().optional(),
    rooms: zod.number().optional(),
    propertyType: zod.string().optional(),
    maxPrice: zod.number().optional(),
    features: zod.array(zod.string()).optional(),
  }),
  suggestions: zod
    .array(zod.string())
    .describe("Search suggestions and alternatives in Persian"),
});

/**
 * @summary List all conversations
 */
export const ListOpenaiConversationsResponseItem = zod.object({
  id: zod.number(),
  title: zod.string(),
  createdAt: zod.coerce.date(),
});
export const ListOpenaiConversationsResponse = zod.array(
  ListOpenaiConversationsResponseItem,
);

/**
 * @summary Create a new conversation
 */
export const CreateOpenaiConversationBody = zod.object({
  title: zod.string(),
});

/**
 * @summary Get conversation with messages
 */
export const GetOpenaiConversationParams = zod.object({
  id: zod.coerce.number(),
});

export const GetOpenaiConversationResponse = zod.object({
  id: zod.number(),
  title: zod.string(),
  createdAt: zod.coerce.date(),
  messages: zod.array(
    zod.object({
      id: zod.number(),
      conversationId: zod.number(),
      role: zod.string(),
      content: zod.string(),
      createdAt: zod.coerce.date(),
    }),
  ),
});

/**
 * @summary Delete a conversation
 */
export const DeleteOpenaiConversationParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary List messages in a conversation
 */
export const ListOpenaiMessagesParams = zod.object({
  id: zod.coerce.number(),
});

export const ListOpenaiMessagesResponseItem = zod.object({
  id: zod.number(),
  conversationId: zod.number(),
  role: zod.string(),
  content: zod.string(),
  createdAt: zod.coerce.date(),
});
export const ListOpenaiMessagesResponse = zod.array(
  ListOpenaiMessagesResponseItem,
);

/**
 * @summary Send a text message and receive a streaming text response
 */
export const SendOpenaiMessageParams = zod.object({
  id: zod.coerce.number(),
});

export const SendOpenaiMessageBody = zod.object({
  content: zod.string(),
});